Bump ruff to 0.16.3 and satisfy its widened defaults
ai-maintainer and keylight run the same tool, and this was two minors
behind them because 0.16 selects rules 0.15 did not. Landing the version
and the code together, since the gate is only green with both.

Mechanical for the most part: percent formatting to f-strings, sorted
imports, explicit parentheses around the implicit concatenations inside
collection literals, `dict()` to `{}`.

Three that are not:

  - `lib/pcbnew_quiet` must import pcbnew before anything else does, since
    it silences the startup wxASSERT by swapping fd 2 across that import.
    The import sort moved it below `kikit` in panelize.py, which imports
    pcbnew itself, so the guard became a no-op. Marked `isort: skip` in
    the three files where the order carries meaning.

  - The closures in tidy-patterns.py bind their loop variable as a default
    argument. Both are called within the iteration that defines them, so
    this is what they already did, spelled so it stays that way.

  - The blind excepts are deliberate and stay, with the reason at each one.
    Narrowing the top-level handler in gen-case.py in particular would
    defeat it: freecadcmd exits 0 on an uncaught exception.

`scripts/lib/*.py` lose their shebangs (none has a __main__ guard) and
the five top-level scripts that lacked the executable bit now carry it,
as the rest of scripts/ already did.

Verified against a pre-change checkout: validate-symmetry, provenance and
firmware, tidy-patterns, tidy-slivers and cleanup-tracks all print the
same output and touch no board, and gen-case writes STEP files with
identical geometry.

+++ scripts/panelize.py
@@ -21,10 +21,13 @@ import argparse
 import os
 import sys
 
-from lib.pcbnew_quiet import pcbnew
-from kikit.panelize import Panel, Origin
-from kikit.units import mm
+# First, and marked so the import sort leaves it there: it silences pcbnew's
+# startup wxASSERT by swapping fd 2 across the import, which only works if
+# nothing has imported pcbnew yet. kikit below imports pcbnew itself.
+from lib.pcbnew_quiet import pcbnew  # isort: skip
 
+from kikit.panelize import Origin, Panel
+from kikit.units import mm
 from lib.provenance import read_stamp, write_stamp
 
 
@@ -169,7 +172,7 @@ def main():
     # KiCad internal units are integer nanometres; mm scales up, so round to int
     # (argparse gives floats) before these reach pcbnew/KiKit geometry calls.
     def nm(v):
-        return int(round(v * mm))
+        return round(v * mm)
 
     os.makedirs(os.path.dirname(os.path.abspath(args.output)), exist_ok=True)
     build(
